Move homepage styling to Home.css
A body css rule in sign.css was really troublesome
when trying to keep the styling different across
the pages. I commented out this rule which
allowed me to incorporate the background image on
the landing page.

Slight code layout edits in App.jsx putting the header
component back into the Router wrapper.